+++ lib/src/controllers/vimeo_playback_controller.dart
@@ -1,13 +1,19 @@
 import 'dart:async';
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_inappwebview/flutter_inappwebview.dart';
+import 'package:flutter/services.dart';
 import 'package:omni_video_player/omni_video_player/controllers/global_playback_controller.dart';
 import 'package:omni_video_player/omni_video_player/controllers/omni_playback_controller.dart';
 import 'package:omni_video_player/omni_video_player/models/omni_video_quality.dart';
 import 'package:omni_video_player/omni_video_player/models/video_player_callbacks.dart';
 import 'package:omni_video_player/omni_video_player/models/video_source_type.dart';
 import 'package:video_player/video_player.dart';
+import 'package:webview_flutter/webview_flutter.dart';
+import 'package:webview_flutter_android/webview_flutter_android.dart';
+import 'package:webview_flutter_wkwebview/webview_flutter_wkwebview.dart';
+import '../../vimeo/vimeo_player_event_handler.dart';
+import '../utils/logger.dart';
 
 class VimeoPlaybackController extends OmniPlaybackController {
   @override
@@ -19,15 +25,26 @@ class VimeoPlaybackController extends OmniPlaybackController {
   @override
   final String? videoDataSource = null;
 
-  InAppWebViewController? _webViewController;
+  late final WebViewController webViewController;
+  late final VimeoPlayerEventHandler _eventHandler;
+  final Completer<void> _initCompleter = Completer();
+
+  String get playerId => 'Vimeo$hashCode';
 
   bool _isPlaying = false;
   bool _isReady = false;
+  bool _isBuffering = false;
   Duration _currentPosition = Duration.zero;
+  Duration _duration = Duration.zero;
 
   @override
-  final Duration duration;
-  Timer? _positionTimer;
+  Duration get duration => _duration;
+
+  set duration(Duration value) {
+    _duration = value;
+    notifyListeners();
+  }
+
   @override
   final Size size;
 
@@ -38,35 +55,45 @@ class VimeoPlaybackController extends OmniPlaybackController {
   bool _isSeeking = false;
   bool _isFullScreen = false;
   bool _hasStarted = true;
-  bool _isBuffering = true;
   bool _hasError = false;
   final GlobalPlaybackController? _globalController;
   double _previousVolume = 1.0;
   final List<VoidCallback> _onReadyQueue = [];
-
-  void _executeOrQueue(VoidCallback action) {
-    if (_isReady) {
-      action();
-    } else {
-      _onReadyQueue.add(action);
-    }
-  }
 
   VimeoPlaybackController._(
     this.videoId,
     this._globalController,
     Duration initialPosition,
     double? initialVolume,
-    this.duration,
     this.size,
     this.callbacks,
   ) {
-    _executeOrQueue(() {
-      seekTo(initialPosition, skipHasPlaybackStarted: true);
-      if (initialVolume != null) {
-        volume = initialVolume;
-      }
-    });
+    _eventHandler = VimeoPlayerEventHandler(this);
+
+    late final PlatformWebViewControllerCreationParams webViewParams;
+    if (WebViewPlatform.instance is WebKitWebViewPlatform) {
+      webViewParams = WebKitWebViewControllerCreationParams(
+        allowsInlineMediaPlayback: true,
+        mediaTypesRequiringUserAction: const <PlaybackMediaTypes>{},
+      );
+    } else {
+      webViewParams = const PlatformWebViewControllerCreationParams();
+    }
+
+    webViewController = WebViewController.fromPlatformCreationParams(
+      webViewParams,
+    )
+      ..setJavaScriptMode(JavaScriptMode.unrestricted)
+      ..addJavaScriptChannel(playerId, onMessageReceived: _eventHandler.call)
+      ..enableZoom(false);
+
+    final webViewPlatform = webViewController.platform;
+    if (webViewPlatform is AndroidWebViewController) {
+      AndroidWebViewController.enableDebugging(false);
+      webViewPlatform.setMediaPlaybackRequiresUserGesture(false);
+    } else if (webViewPlatform is WebKitWebViewController) {
+      webViewPlatform.setAllowsBackForwardNavigationGestures(false);
+    }
   }
 
   /// Creates and initializes a new [OmniPlaybackController] instance.
@@ -75,7 +102,6 @@ class VimeoPlaybackController extends OmniPlaybackController {
     required GlobalPlaybackController? globalController,
     required Duration initialPosition,
     double? initialVolume,
-    required Duration duration,
     required Size size,
     required VideoPlayerCallbacks callbacks,
   }) {
@@ -84,9 +110,40 @@ class VimeoPlaybackController extends OmniPlaybackController {
       globalController,
       initialPosition,
       initialVolume,
-      duration,
       size,
       callbacks,
+    );
+  }
+
+  Future<void> init() async {
+    final platform = kIsWeb ? 'web' : defaultTargetPlatform.name.toLowerCase();
+
+    await webViewController.loadHtmlString(
+      await _buildPlayerHTML(<String, String>{
+        'videoId': videoId,
+        'platform': platform,
+        'playerId': playerId,
+      }),
+      baseUrl: kIsWeb ? Uri.base.origin : "https://player.vimeo.com",
+    );
+
+    if (!_initCompleter.isCompleted) _initCompleter.complete();
+  }
+
+  @override
+  Future<void> dispose() async {
+    super.dispose();
+    await webViewController.removeJavaScriptChannel(playerId);
+  }
+
+  Future<String> _buildPlayerHTML(Map<String, String> data) async {
+    final playerHtml = await rootBundle.loadString(
+      'packages/omni_video_player/assets/vimeo_player.html',
+    );
+
+    return playerHtml.replaceAllMapped(
+      RegExp(r'<<([a-zA-Z]+)>>'),
+      (m) => data[m.group(1)] ?? m.group(0)!,
     );
   }
 
@@ -180,7 +237,8 @@ class VimeoPlaybackController extends OmniPlaybackController {
   }
 
   @override
-  bool get isFinished => duration == currentPosition;
+  bool get isFinished =>
+      currentPosition >= duration && duration != Duration.zero;
 
   @override
   int get rotationCorrection => 0;
@@ -295,53 +353,13 @@ class VimeoPlaybackController extends OmniPlaybackController {
   }
 
   Future<void> _evaluate(String js) async {
-    if (_webViewController == null) return;
+    await _initCompleter.future;
+
     try {
-      await _webViewController!.evaluateJavascript(source: js);
+      await webViewController.runJavaScript(js);
     } catch (e) {
-      debugPrint('Error evaluating JS: $js\n$e');
+      logger.w('Error evaluating JS: $js\n$e');
     }
-  }
-
-  DateTime? _lastTickTime;
-
-  void startPositionTimer() {
-    _positionTimer?.cancel();
-    _lastTickTime = DateTime.now();
-
-    _positionTimer = Timer.periodic(Duration(milliseconds: 500), (_) {
-      if (_isPlaying) {
-        final now = DateTime.now();
-        final elapsed = now.difference(_lastTickTime!);
-        _lastTickTime = now;
-
-        final newPosition = _currentPosition + elapsed;
-
-        if (newPosition > duration) {
-          _currentPosition = duration;
-          _positionTimer?.cancel();
-          pause();
-        } else {
-          _currentPosition = newPosition;
-        }
-
-        notifyListeners();
-      }
-    });
-  }
-
-  void stopPositionTimer() {
-    _positionTimer?.cancel();
-  }
-
-  void setWebViewController(InAppWebViewController? controller) {
-    _webViewController = controller;
-  }
-
-  @override
-  void dispose() {
-    _positionTimer?.cancel();
-    super.dispose();
   }
 
   @override